make ispyb callback base class
